Fix workflow transition to status reopened always checked
In manage_config_workflow_page.php, the code was always showing the
'reopened' states as checked regardless of the actual values stored in
the database.

This was due to storing the 'reopened' label within the array of values
used to build the table to display, which consequently were always set.

The code that initializes the labels has been removed from function
parse_workflow(), and the logic is now handled directly in show_flag(),
with globally initialized variables for better performance.

Fixes #12129