Improved get_status_option_list() function
Previous behavior to return all available values for the status enum
when the workflow is not set for the given status, could potentially
result in a security issue (letting the user bypass the workflow). The
function now returns an empty array in this case, leaving it up to the
caller to handle a possibly empty list.

Simplified and optimized the test in the foreach loop.

Do not add 'closed' status to the list if it is the current status.

Added PHPdoc function header.

Affects #11661, #14156