Add 'Close' button for bugs
This commit introduces a new 'Close' button in the View Issue Details
page, making it easier and more intuitive for reporters to close their
own issues (when $g_allow_reporter_close = ON).

It also improves the access level verifications, simplifies the code and
makes it more readable by using standard access check functions.

This is an improvement over the functionality implemented in #11502 not
only in terms of usability, but also for security as in some specific
cases users could actually bypass the workflow.

Fixes #14156