Use BugData object as function param for several functions
Commit 085ab1697e93ab4c9384a8268189000583452207 introduced an undefined
variable warning, due to removal of a call to bug_get_field function to
retrieve the project_id. This could actually cause the incorrect status
to be retrieved in html_button_bug_reopen() and html_button_bug_close(),
potentially allowing users to close or reopen issues when they are not
entitled to.

Instead of adding the function call back as it was, the code was
modified to use a BugData object instead of a bug id as parameter for
several functions in access_api.php and html_api.php, resulting in
simpler and more efficient code thanks to fewer calls to bug_get() and
bug_get_field().

Fixes #14191, Affects #14156