Add filter in Manage Users page to show/hide disabled users
This is a port of the following 1.2.x commits:

   - a21374af650c06a27bf9a930edbcbef9b41989ac,

Original patch written by David Newcomb.

Fixed issue with new filter not specified on the multi-page navigation
links. Also made minor cosmetic changes: whitespace, coding guidelines.

   - 129a8e5e93fda7537884bea371b794bb44c13c0c

Simplify code by using a for loop to print the column headers instead of
manually building the table cells one by one. Note that this actually
changes the label of the 'Protected' column from the padlock icon to a
text description.

Rename 'hide' variable to 'hide_inactive'. With the addition of the new
filter option 'show_disabled', using just 'hide' was potentially
confusing.

   - e28fa43081f8365fafbc0c2331938032a3fe89a4

The filter in managed_user_page.php is not persistent

The filter values stored in the cookie whenever the criteria change,
were restored to the form variables instead of the sanitized ($c_)
variables, and were therefore not taken into account when building the
page.

Fixes #11726, #14216