Code cleanup in Manage User page
Simplify code by using a for loop to print the column headers instead of
manually building the table cells one by one. Note that this actually
changes the label of the 'Protected' column from the padlock icon to a
text description.

Rename 'hide' variable to 'hide_inactive'. With the addition of the new
filter option 'show_disabled', using just 'hide' was potentially
confusing.