Fix #13304: Error handler causes XML Parsing Errors
The standard error handler was generating invalid HTML, due to not
outputting previously issued page headers.

To fix this issue, we retrieve the output buffer's contents, and retain
all html up to the </head> tag (not included).

A few layout changes were applied as well, to make better use of
screen real estate so that the error page is easier to read, especially
when $g_show_detailed_errors is ON.