Record reporter+submitted date in extended info first revision history
When updating any of the "extended info" fields (Description, Steps To
Reproduce, Additional Information), the history of modifications is
saved in the mantis_bug_revision_table. When the first change is made
since issue creation, an additional, "initial" revision record is
created to keep track of the original value.

Prior to this, the record was saved with incorrect author and timestamp
information: current user and date of modification, instead of the bug's
reporter and date of submission.

Fixes #14272