Fix PostgreSQL error when adding project/subproject (schema change)
The error is caused by columns mantis_project_table.inherit_global and
mantis_project_hierarchy_table.inherit_parent, which are defined as
unsigned int in schema.php, but treated as boolean in the code. This is
a problem with PostgreSQL due to strict type checking, but not on MySQL
as type cast is done automatically.

Changes in schema.php the type for the 2 fields project.inherit_global
and project_hierarchy.inherit_parent from Integer to Boolean

Fixes #14385