Fix PostgreSQL error when updating a project
Release 1.2.11 (see commit b8d4b5039598248d0b0c78619450c51d4dc98df2 and
issue #14288) introduced a regression preventing the user from updating
an existing project.

This commit is a workaround for the problem (sending an int to the DB
instead of a bool if using PostgreSQL), as fixing the root cause would
require a schema change which is not possible in 1.2.x.

Fixes #14385