Fix #11687: Bugs with attachments that are moved will lose attachments
Backporting 08c027af59b88a6934377d71d19a86edfd7d12dd to 1.2.x branch.

When a bug is logged and assigned to a project which has a project path
assigned, and later a mantis "administrator" moves the bug from the
initial project to a new project, the attachments are lost.

Signed-off-by: Damien Regad <damien.regad@merckgroup.com>