Add param to bug_resolve() to allow "resolution" to custom status
Function bug_resolve() sets the status to bug_resolved_status_threshold.
This behavior is fine with MantisBT's default status enum, but is
causing issues when using a custom target status that is
>= $g_bug_resolved_status_threshold & < $g_bug_closed_status_threshold.

In this case, when changing status using the button on view.php, the
target status selected by the user is overriden by the value of
bug_resolved_status_threshold, resulting in incorrect behavior. The only
way to effectively reach the custom state is by editing the issue.

To address this, a new optional parameter $p_status is added to
bug_resolve(). To make the function call more logical, the signature was
modified: new parameter inserted in 3rd position after $p_resolution.

The function's phpdoc header has been completed.

Fixes #14443