Cloning issues with attachments fails after upload folder changed
With attachments stored on disk, if the project's upload directory
changes and the files are moved to the new location, it is no longer
possible to clone issues.

The file_copy_attachments() function was modified to use the same logic
as elsewhere in MantisBT, i.e. rely on file_normalize_attachment_path()
to determine the attachment's path.

Fixes #14718