Force validation of upload path if file_upload_method != DATABASE
Prior to this, the upload path would not be checked if it was empty
(ie. defaulting to $g_absolute_path_default_upload_folder),
which could cause issues if that directory was missing or read-only.

Checking this early allows catching any errors before users perform
actual operations on file attachments.

Fixes #14723