Fixes #14756: mismatched types passed on to check_selected()
When trying to report an issue in a setup having subprojects, and no
specific project is selected (current project = 'All projects'), an
error is triggered in check_selected() due to mismatched types.

Statically casting the project_id as string to solve the problem.

Thanks to phyllisl for detecting and proposing a solution to this issue.