Fix #14842: Incorrect matching of versions in print_version_option_list
When comparing versions to determine which ones have already been added
to the select list, a loose comparison is performed, so '1' matches
'1.0'.

This causes issues when editing a bug having a version currently
obsolete matching another, previously displayed version. In that case
the obsolete version is not added to the selection list, causing data
loss when the bug is saved (see issue #13096).