Fix and improve timezone verification in admin checks
admin/check.php now properly detects issues with timezone settings, in
line with behavior of date_default_timezone_get() in PHP >= 5.3 and
provides better information to the admin in case of errors.

It also fixes an issue with PHP 5.1, by only calling function
timezone_identifiers_list() when it actually exists, and returns a
warning that it is unable to check timezone if it does not. Thanks to
Tim Jackson for detecting and providing the initial patch for this bug.

Fixes #14009