Manage columns include: rename form and remove multipart encoding
The form in manage_columns_inc.php was apparently copy-pasted from
bug_report_page.php; it was named "report_bug_form" and an optional
multipart encoding if file upload is enabled, which does not make sense
in this context.

This code-cleanup commit removes the enctype and renames the form.

Fixes #15280

Signed-off-by: Damien Regad <damien.regad@merckgroup.com>