Removing current user from list built by print_user_option_list()
Since there is a [myself] value in the list, the current user should be
removed. This was done for ALL_PROJECTS with the fix for issue #10130,
but the single-project code still listed the current user.

By moving the code to remove the current user to the foreach loop that
builds the sort arrays, we ensure that it is removed in both cases.