Remove call to realpath() when setting BASE_PATH in core.php
Per PHP documentation [1], since version 4.0.2 __FILE__ always contains
an absolute path with symlinks resolved; considering that MantisBT's
minimum requirement is PHP 5.1 we therefore do not need to call
realpath().

This avoids problems with the function returning FALSE when the running
script does not have executable permissions on all directories in the
hierarchy, which can happen e.g. with shared hosting on Windows.

Fixes #15357

[1] http://php.net/language.constants.predefined.php