Modify file_is_name_unique() API to work for project files also
Previously this function would only work for bug attachments, and
another API function project_file_is_name_unique() would take care of
projects. The latter was removed as it was not used, and this one
amended in case it's ever required to check for uniqueness of a project
file.

It also makes sense to regroup file-related functions in file_api.php.

Issue #15572