Config report: retrieval of saved project filter does not work
When retrieving a saved filter from the cookie in adm_config_report.php
the saved project id is not reflected in the filter's selection list:
'All Projects' is always selected instead of the actual project. This
value should only be picked as default when the project id does not
exist.

This is due to a missing negation in the check for project's existence.

Fixes  #15691