Fix regression in filter_api.php
Commit c9bc0646d69b2e39cd5e48320be82a83d8e39ebc introduced a regression,
whereby the filter would exclude issues without bugnotes when using text
search.

This was caused by removal of left outer join between bugnote_table and
bugnote_text_table, which was in fact required because the join clause
fails to return any record when the bugnote_text_id is null (which is
the case when there are no bugnotes).

Thanks to user Kitzberger for reporting this and atrol for providing the
steps to reproduce.