Changed position and label of 'Changed(Hrs)' filter
This setting is actually not a filter but a display option that
just highlights the last updated date. The field has been moved next to
'Sort by', and its label changed to reflect its real function.

Fixes #7737

This is a port of the original 1.2.x commits from the pull request
(squashed).

Signed-off-by: Damien Regad <dregad@mantisbt.org>