Disable selection of projects in which user can't report issues
When the current project is 'All Projects' and user clicks on 'Report
Issue', login_select_proj_page.php presents them with a list of projects,
which includes those in which the user is not allowed to report issues.
If one of these projects is selected, an 'Access Denied' error occurs.

This commit makes the functionality more user-friendly by disabling
these projects in the list, so users can't select them.

To implement this, a new optional parameter was added to functions
print_project_option_list() and print_subproject_option_list().

Fixes #16024

Conflicts:
	core/print_api.php