Can't copy columns to/from subproject when parent is current
In the manage columns page, it is not possible to copy the columns to or
from a subproject, when the current project is the parent project,
because the selection list completely removes the whole subtree instead
of just the current project.

In addition, we now also don't print 'All Projects' anymore when the
current project is the same.

Fixes #16029

Conflicts:
	core/print_api.php