Follow-up fix for custom field names translations
In the case where we have multiple custom strings localized to the same
string, commit c174d88fc4f6a92e5bff0f7660688f1757c4cdaf would make it
impossible to distinguish them.

While this should not be a major issue in the bug history (as it doesn't
make sense to have more than one fields with a given localized name
within a given project), it is important that in custom field management
as well as project management pages custom fields are uniquely
identified with their "system" name to ensure there is no confusion.

This commit adds a new API function custom_field_get_display_name() that
will append the localized name to the field name, and which is called
from the above-mentioned pages.

Issue #12470