Reintroduce form_action_self() api
This is a partial revert of 38b0212651d368b85858a3820ddbc0150ae32a42.

There are some cases, specifically in the manage_config_* pages, where
form_action_self() is not used directly as a form's action, but passed
to another script to be used for redirections.

Using '' instead of $_SERVER['PHP_SELF'] in this context results in
incorrect redirection (i.e. Mantis will redirect to the user's default
page).

Fixes #16410