Workflow thresholds page: fix display of 'who can alter'
When the current project is 'All projects' and the admin changes the
'Who can alter value' threshold (e.g. to 'manager'), the changes are saved
in the config table as expected, but not displayed back properly -
mantis shows the default value of 'administrator' (admin_site_threshold)
instead, giving the impression that the settings have not been saved;
another update will then overwrite the previously saved threshold back
to default.

The projects list in config_get_access() was not properly initialized,
set to an empty array when current project is ALL_PROJECTS.

Fixes #10071