Workflow thresholds page: add color highlight for 'who can alter'
Normally on this page, all changes versus system / global defaults are
highlighted in color. This did not work for the access level threshold
of 'Who can alter this value', which are never color-coded regardless of
the actual value.

This commit introduces a new function print_who_can_change() to remove
code duplication.

Fixes #16411