Fixes #16416: Improve first login experience by auto-redirecting to create project page.
If MantisBT is pre-installed, the user lands in My View Page without knowing what to do.
We should handle the "first login" experience in MantisBT itself rather than the installation
and provide a helpful getting started over time. For now, we should start with auto-redirecting
to create project page if user is administrator and there aren't at least 1 project in the database.

The cost of this check is one cheap query on the projects table that returns 0 or 1 entries
which is only run at login of an administrator.