Individually number each schema upgrade step
This is basically a redo of ad7009ca4a6810ddd559a8613c8a5b05e2fc5c6b
which was reverted before rebasing the Oracle branch.

The difference is that instead of adding a comment before each upgrade,
we are now specifying the index as the array element is initialized.
Advantages: self-documented code, easier maintenance.

Use the command below to insert a space-padded schema number for each
upgrade step:
$ perl -lne 'if(/^\$upgrade/) {$s=sprintf("%0 s",$n++); s/\[\]/[$s]/} print $_' schema.php >x && mv x schema.php