Fix layout for bug note avatar and move it to the left side
The text next to the avatar in the bug notes was being pushed under the avatar.  This change fixes that and moves the avatar to the left of the text which is more standard with other conventions.  As we add avatars elsewhere in MantisBT, they will also be on the left side.