Proper processing of unhandled errors
Sometimes the execution of a test itself can trigger an error, and in
some cases the test can be successful despite of that (e.g. a PHP notice
or use of deprecated functionality). The code was not handling these
situations properly when $g_show_all = true, causing the test to
continue silently instead of reporting the problem.

Added a case for E_DEPRECATED and clarified description for unhandled
error types in check_print_error_rows(). Also made sure the error is
properly captured  when hiding passed tests and/or verbose error
messages are not shown.

Fixes #16890