Allow Reporters to Clone issues
Previously, users needed to have update_bug_threshold to be able to
clone issues, since the creation of a relationship with the parent bug
requires to update it.

This patch lets Reporters clone issues, but disables the "Relationship
with the parent issue" field in bug_report_page.php when the user does
not have updater access to the parent bug.

It also modifies the access level check performed against the parent
bug's project in bug_report.php, which is now only needed when a
relationship with it has been specified.

Fixes #11290 (implements option 1)